backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.14
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
greenlet==3.2.3
h11==0.16.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.4
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pillow==11.3.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.173
ultralytics-thop==2.0.15
urllib3==2.5.0
uvicorn==0.35.0
